Safari-609.1.11.1/LayoutTests/webgl/2.0.0/conformance2/reading/read-pixels-pack-parameters-expected.txt

CONSOLE MESSAGE: line 159: TypeError: Argument 7 ('pixels') to WebGL2RenderingContext.readPixels must be an instance of ArrayBufferView
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/reading/read-pixels-pack-parameters.html
[ 1: PASS ] context exists
[ 2: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 3: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 4: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 5: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 6: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 7: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 8: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 9: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 10: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 11: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 12: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 13: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 14: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 15: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 16: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 17: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 18: FAIL ] getError expected: NO_ERROR. Was INVALID_ENUM : readPixels should succeed
[ 19: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 20: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 21: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 22: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 23: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 24: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 25: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 26: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 27: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 28: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 29: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 30: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 31: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 32: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 33: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 34: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 35: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 36: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 37: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 38: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 39: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 40: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 41: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 42: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 43: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 44: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 45: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 46: PASS ] getError was expected value: INVALID_OPERATION : Invalid pack params combination
[ 47: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 48: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 49: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 50: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 51: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 52: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 53: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 54: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 55: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 56: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 57: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 58: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 59: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 60: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 61: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 62: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 63: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 64: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 65: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 66: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 67: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 68: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 69: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 70: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 71: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 72: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 73: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 74: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 75: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 76: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 77: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 78: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 79: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 80: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 81: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 82: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 83: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 84: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 85: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 86: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 87: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 88: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 89: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 90: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : readPixels should succeed
[ 91: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 92: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 93: FAIL ] Padding byte 4 of row 0 changed: expected 1, got 2
[ 94: FAIL ] last pixel of row 1: expected [2,200,102,255], got [134,87,234,255]
[ 95: FAIL ] Padding byte 0 of row 1 changed: expected 1, got 134
[ 96: FAIL ] first pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 97: FAIL ] last pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 98: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 99: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 100: FAIL ] Padding byte 8 of row 0 changed: expected 1, got 2
[ 101: FAIL ] first pixel of row 1: expected [1,2,3,4], got [2,200,102,255]
[ 102: FAIL ] last pixel of row 1: expected [2,200,102,255], got [1,2,3,4]
[ 103: FAIL ] Padding byte 0 of row 1 changed: expected 1, got 134
[ 104: FAIL ] last pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 105: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 106: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 107: FAIL ] Padding byte 4 of row 0 changed: expected 1, got 249
[ 108: FAIL ] last pixel of row 1: expected [249,102,0,255], got [2,200,102,255]
[ 109: FAIL ] Padding byte 0 of row 1 changed: expected 1, got 2
[ 110: FAIL ] first pixel of row 2: expected [2,200,102,255], got [1,2,3,4]
[ 111: FAIL ] last pixel of row 2: expected [2,200,102,255], got [1,2,3,4]
[ 112: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 113: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 114: FAIL ] Padding byte 8 of row 0 changed: expected 1, got 249
[ 115: FAIL ] first pixel of row 1: expected [1,2,3,4], got [249,102,0,255]
[ 116: FAIL ] last pixel of row 1: expected [249,102,0,255], got [1,2,3,4]
[ 117: FAIL ] Padding byte 0 of row 1 changed: expected 1, got 2
[ 118: FAIL ] last pixel of row 2: expected [2,200,102,255], got [1,2,3,4]
[ 119: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 120: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 121: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 122: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 123: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 124: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 125: FAIL ] Padding byte 4 of row 0 changed: expected 1, got 2
[ 126: FAIL ] first pixel of row 1: expected [2,200,102,255], got [1,2,3,4]
[ 127: FAIL ] last pixel of row 1: expected [1,2,3,4], got [134,87,234,255]
[ 128: FAIL ] Padding byte 0 of row 1 changed: expected 1, got 134
[ 129: FAIL ] first pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 130: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 131: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 132: FAIL ] Padding byte 4 of row 0 changed: expected 1, got 99
[ 133: FAIL ] last pixel of row 1: expected [99,5,76,255], got [1,2,3,4]
[ 134: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 135: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 136: FAIL ] Padding byte 4 of row 0 changed: expected 1, got 99
[ 137: FAIL ] first pixel of row 1: expected [99,5,76,255], got [1,2,3,4]
[ 138: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 139: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 140: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 141: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 142: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 143: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : Invalid pack params combination
[ 144: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 145: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : Invalid pack params combination
[ 146: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 147: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 148: FAIL ] skipped bytes changed at index 0: expected 1 got 249
[ 149: FAIL ] first pixel of row 0: expected [249,102,0,255], got [1,2,3,4]
[ 150: FAIL ] last pixel of row 0: expected [249,102,0,255], got [1,2,3,4]
[ 151: FAIL ] first pixel of row 1: expected [2,200,102,255], got [1,2,3,4]
[ 152: FAIL ] last pixel of row 1: expected [2,200,102,255], got [1,2,3,4]
[ 153: FAIL ] first pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 154: FAIL ] last pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 155: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 156: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : Invalid pack params combination
[ 157: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 158: FAIL ] getError expected: INVALID_OPERATION. Was NO_ERROR : Invalid pack params combination
[ 159: FAIL ] getError expected: INVALID_OPERATION. Was INVALID_ENUM : buffer too small
[ 160: PASS ] getError was expected value: NO_ERROR : readPixels should succeed
[ 161: FAIL ] skipped bytes changed at index 0: expected 1 got 249
[ 162: FAIL ] first pixel of row 0: expected [249,102,0,255], got [1,2,3,4]
[ 163: FAIL ] last pixel of row 0: expected [249,102,0,255], got [1,2,3,4]
[ 164: FAIL ] first pixel of row 1: expected [2,200,102,255], got [1,2,3,4]
[ 165: FAIL ] last pixel of row 1: expected [2,200,102,255], got [1,2,3,4]
[ 166: FAIL ] first pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 167: FAIL ] last pixel of row 2: expected [134,87,234,255], got [1,2,3,4]
[ 168: FAIL ] successfullyParsed should be true (of type boolean). Was undefined (of type undefined).
[ FAIL ] 142 failures reported
